Implemented User input
Instead of hardcoding the positions of pieces, i created functions to dynamically update the board after each input from user. Used the stack data structure from until to create a undo function. Still have not implement where it is either a win, lost or tie, But as of right now it works as a human vs human.